+++ internal/video/font/main_test.go
@@ -1,0 +1,11 @@
+package font
+
+import (
+	"testing"
+
+	"basics/testutils"
+)
+
+func TestMain(m *testing.M) {
+	testutils.RunWithAssertTracking(m)
+}